+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - `stairs` Stairstep graph.
 - `fill` 2-D patch.
 - `subsref` Subscripted reference.
+- Intel compiler support.
 
 ### Changed